Open json with utf-8 encoding
Without this, it is possible to get a UnicodeDecodeError because the
file is opened as ascii.

I am not sure why the file is opened as ascii in Python 3. I am also
unsure why it would happen now if it hadn't happened before. However,
whatever the cause, being explicit fixes it.